requirements.txt
Django==5.2
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.10
pymongo==3.11.4
django-environ==0.12.0
torch==2.2.0
torchaudio==2.2.0
clap-pytorch==1.0.0
numpy==2.2.5 